djangoapp/requirements.txt
Django>=2.1.2
django-import-export
